# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Laravel application",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a list of articles",
                "description": "Retrieve a list of articles with optional filters for keyword, date, category, and source. Results are paginated.",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search for articles by keyword in the title",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "euro"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter articles by publication date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-10"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "sports"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "New York Times"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Breaking News: Euro Hits All-Time High"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "The Euro reached a record high today."
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "datetime",
                                                        "example": "2024-12-10T10:00:00Z"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Finance"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "New York Times"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles?page=1"
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles?page=5"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles?page=2"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a specific article by ID",
                "description": "Retrieve the details of a specific article by its ID.",
                "operationId": "getArticleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the article to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Breaking News: Euro Hits All-Time High"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "The Euro reached a record high today."
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "datetime",
                                            "example": "2024-12-10T10:00:00Z"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Finance"
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "New York Times"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}